Added ability to read raw pcm.
This facilitates using lame to stream straight to waveform:

% lame --decode -t file.mp3 - | waveform --raw --peak

No need for intermediate file.  Note that waveform will collect the
entire data set in memory before processing as it has to know the size
before it can proceed.  However, it needs very little more memory than
the pcm data so it shouldn't be a problem on modern systems with normal
(~50MB) files.